+++ src/content/etherscan/feat-scripts/fund-flow.tsx
@@ -1,5 +1,7 @@
 import { createRoot } from 'react-dom/client'
 
+import { FUND_FLOW_UNSUPPORTED_LIST } from '@common/constants'
+
 import { FundFlowBtn } from '../components'
 
 const setBtn = (txHashEl: HTMLElement, chain: string, mainAddress: string) => {
@@ -15,6 +17,8 @@ const setBtn = (txHashEl: HTMLElement, chain: string, mainAddress: string) => {
 
 /** fund flow */
 const genFundFlow = async (chain: string) => {
+  if (FUND_FLOW_UNSUPPORTED_LIST.includes(chain)) return
+
   const mainAddress =
     document.querySelector<HTMLElement>('#mainaddress')?.innerText
   if (!mainAddress) return